# entitlement-service/entitlement-service-openapi.yml
openapi: 3.0.3
info:
  title: Entitlement Service API
  description: |
    NextGen Workflow Entitlement Service - Central authority for authentication, authorization, and user management.
    
    This service provides:
    - Session-based authentication
    - Cerbos policy-based authorization
    - User and role management
    - Business application management
    
    **Authentication**: All endpoints except `/api/auth/login` require session-based authentication via `X-Session-Id` header.
    
    **Authorization**: Fine-grained access control is provided through Cerbos policy engine integration.
  version: 1.0.0
  contact:
    name: NextGen Workflow Team
servers:
  - url: http://localhost:8081
    description: Development server
...
paths:
...
  /api/entitlements/users:
    get:
      tags:
        - Users
      summary: Get all users
      description: Retrieve all users in the system
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'

    post:
      tags:
        - Users
      summary: Create user
      description: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid user data
  /api/entitlements/users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      description: Retrieve a specific user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User ID (UUID format)
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found

    put:
      tags:
        - Users
      summary: Update user
      description: Update an existing user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User ID (UUID format)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found

    delete:
      tags:
        - Users
      summary: Delete user
      description: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User ID (UUID format)
      responses:
        '204':
          description: User deleted successfully
        '404':
          description: User not found
  /api/entitlements/users/username/{username}:
    get:
      tags:
        - Users
      summary: Get user by username
      description: Retrieve a user by username
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
          description: Username
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
  /api/entitlements/users/email/{email}:
    get:
      tags:
        - Users
      summary: Get user by email
      description: Retrieve a user by email address
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
          description: Email address
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
...
  /api/entitlements/users/{id}/activate:
    patch:
      tags:
        - Users
      summary: Activate user
      description: Activate a user account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User ID (UUID format)
      responses:
        '200':
          description: User activated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
  /api/entitlements/users/{id}/deactivate:
    patch:
      tags:
        - Users
      summary: Deactivate user
      description: Deactivate a user account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User ID (UUID format)
      responses:
        '200':
          description: User deactivated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found

  # Business Application endpoints
...
  /api/entitlements/user-business-app-roles/user/{userId}:
    get:
      tags:
        - User Business App Roles
      summary: Get user roles by user ID
      description: Retrieve all role assignments for a specific user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User ID (UUID format)
      responses:
        '200':
          description: User roles retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserBusinessAppRole'
  /api/entitlements/user-business-app-roles/user/{userId}/active:
    get:
      tags:
        - User Business App Roles
      summary: Get active user roles by user ID
      description: Retrieve all active role assignments for a specific user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User ID (UUID format)
      responses:
        '200':
          description: Active user roles retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserBusinessAppRole'
...
  /api/entitlements/user-departments/user/{userId}:
    get:
      tags:
        - User Department Assignments
      summary: Get user departments by user ID
      description: Retrieve all department assignments for a specific user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User ID (UUID format)
      responses:
        '200':
          description: User departments retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDepartment'
  /api/entitlements/user-departments/user/{userId}/active:
    get:
      tags:
        - User Department Assignments
      summary: Get active user departments by user ID
      description: Retrieve all active department assignments for a specific user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User ID (UUID format)
      responses:
        '200':
          description: Active user departments retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDepartment'
...
components:
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440001"
        username:
          type: string
          example: "alice.intake"
        email:
          type: string
          example: "alice.intake@company.com"
        firstName:
          type: string
          example: "Alice"
        lastName:
          type: string
          example: "Smith"
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
        attributes:
          type: object
          additionalProperties: true
          example:
            department: "HR"
            role: "INTAKE_ANALYST"
    BusinessApplication:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        businessAppName:
          type: string
          example: "onecms"
        description:
          type: string
          example: "OneCMS Case Management System"
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
        metadata:
          type: object
          additionalProperties: true
          example:
            version: "1.0.0"
            owner: "Case Management Team"
    BusinessAppRole:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        businessApplication:
          $ref: '#/components/schemas/BusinessApplication'
        roleName:
          type: string
          example: "INTAKE_ANALYST"
        roleDisplayName:
          type: string
          example: "Intake Analyst"
        description:
          type: string
          example: "Analyst responsible for case intake and initial review"
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
        metadata:
          type: object
          additionalProperties: true
          example:
            permissions: ["READ", "CREATE"]
            queue: "intake-analyst-queue"
    UserBusinessAppRole:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        userId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440001"
        businessAppRole:
          $ref: '#/components/schemas/BusinessAppRole'
        isActive:
          type: boolean
          example: true
...
    Department:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        departmentName:
          type: string
          example: "Human Resources"
        departmentCode:
          type: string
          example: "HR"
        description:
          type: string
          example: "Human Resources Department"
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
    UserDepartment:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        userId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440001"
        department:
          $ref: '#/components/schemas/Department'
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"